+++ NEWS.md
@@ -1,6 +1,7 @@
 # labelled 1.0.0-dev
 
 * bug fix: argument drop_unused_labels could now be used with `to_factor.data.frame()`
+* new labels argument for `to_labelled` method when applied to a factor
 
 # labelled 1.0.0